Linux: Move interop code into an Interop/Linux folder
