+++ SeeShark/Interop/Linux/Libc/V4l2InputFormat.cs
@@ -4,7 +4,7 @@
 
 using System.Runtime.Versioning;
 
-namespace SeeShark.Interop.Libc;
+namespace SeeShark.Interop.Linux.Libc;
 
 // Obtained from Linux kernel source code
 // https://github.com/torvalds/linux/blob/7cdabafc001202de9984f22c973305f424e0a8b7/include/uapi/linux/videodev2.h#L541-L831
